FEAT: Split crypto wrapper and add wrapper for pem file
