+++ src/wrappers/cryptosignwrapper.h
@@ -1,38 +1,23 @@
 #ifndef ERD_WRAPPER_CRYPTO_SIGN_H
 #define ERD_WRAPPER_CRYPTO_SIGN_H
 
-#include <sodium.h>
 #include <cstring>
 
+#include "internal/internal.h"
 #include "strchr.h"
 
-#define PUBLIC_KEY_BYTES crypto_sign_PUBLICKEYBYTES
-#define SECRET_KEY_BYTES crypto_sign_SECRETKEYBYTES
-#define SEED_BYTES crypto_sign_SEEDBYTES
-#define SIGN_BYTES crypto_sign_BYTES
+#define PUBLIC_KEY_BYTES 32U
+#define SECRET_KEY_BYTES 64U
+#define SEED_BYTES 32U
+#define SIGN_BYTES 64U
 
 namespace wrapper
 {
 namespace crypto
 {
-std::string getSignature(bytes const &secretKey, std::string const &message)
-{
-    unsigned long long const msgLength = message.length();
-    auto *msg = new unsigned char[msgLength + 1ULL];
-    strcpy((char *) msg, message.c_str());
+std::string getSignature(bytes const &secretKey, std::string const &message);
 
-    unsigned char sig[SIGN_BYTES];
-    unsigned char sk[SECRET_KEY_BYTES];
-
-    std::copy(secretKey.begin(), secretKey.end(), sk);
-
-    unsigned long long signLength;
-    crypto_sign_detached(sig, &signLength, msg, msgLength, sk);
-    std::string ret = util::uCharToStr(sig, signLength);
-
-    delete[] msg;
-    return ret;
-}
+bytes getSecretKey(bytes const &publicKey, bytes const &seed);
 }
 }
